[fix] : 게시물 API(migration) 로직 및 테스트 코드 수정
- 로그인 상태로 게시글 생성, 삭제, 수정하므로 authorId를 findOne하는 로직 삭제
- 이전에 게시물 조회 중 쿼리에 keyword 적용안됨 ->  typeorm 잘못된 코드 수정
- 게시물 수정에 update문에 authorId 조건 추가

+++ nestjs-migration/src/post/post.service.spec.ts
@@ -1,22 +1,26 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { PostService } from "./post.service";
-import { PostRepository } from "./post.repository";
 import { CreatePostDto } from "./dto/create-post.dto";
-import { UserRepository } from "../user/user.repository";
-import { LogRepository } from "../log/log.repository";
-import { DataSource } from "typeorm";
+import { DataSource, Repository } from "typeorm";
 import { Post } from "./entities/post.entity";
 import { UpdatePostDto } from "./dto/update-post.dto";
 import { User } from "../user/entities/user.entity";
 import { ServerError } from "../common/exceptions/server-error.exception";
+import { getRepositoryToken } from "@nestjs/typeorm";
+import { Log } from "../log/entity/log.entity";
+import { PostRepository } from "./post.repository";
+import { LogRepository } from "../log/log.repository";
 
 describe("PostService", () => {
+
 	let postService: PostService;
 	let postRepository: PostRepository;
-	let userRepository: UserRepository;
 	let logRepository: LogRepository;
 	let dataSource: DataSource;
-	let mockQueryRunner: any;
+
+	let mockUserId: number;
+	let mockUser: Partial<User>;
+	let mockPostId: number;
 
 	const mockPostRepository = {
 		getPostHeaders: jest.fn(),
@@ -28,19 +32,12 @@ describe("PostService", () => {
 		update: jest.fn(),
 		findOne: jest.fn(),
 	};
-	const mockUserRepository = {
-		findOne: jest.fn(),
-	};
 	const mockDataSource = {
 		createQueryRunner: jest.fn(),
 	};
 	const mockLogRepository = {
 		save: jest.fn(),
 	};
-
-	let mockUserId: number;
-	let mockUser: Partial<User>;
-	let mockPostId: number;
 
 	beforeEach(async () => {
 		const module: TestingModule = await Test.createTestingModule({
@@ -49,10 +46,6 @@ describe("PostService", () => {
 				{
 					provide: PostRepository,
 					useValue: mockPostRepository,
-				},
-				{
-					provide: UserRepository,
-					useValue: mockUserRepository,
 				},
 				{
 					provide: LogRepository,
@@ -67,19 +60,19 @@ describe("PostService", () => {
 
 		postService = module.get<PostService>(PostService);
 		postRepository = module.get<PostRepository>(PostRepository);
-		userRepository = module.get<UserRepository>(UserRepository);
-		dataSource = module.get<DataSource>(DataSource);
 		logRepository = module.get<LogRepository>(LogRepository);
+		dataSource = module.get<DataSource>(DataSource);
 
 		mockUser = {
 			id: 1,
+			nickname: "Author",
 		};
-		mockUserId = mockUser.id;
+		mockUserId = mockUser.id
 		mockPostId = 2;
 	});
 
 	describe("createPost", () => {
-		mockQueryRunner = {
+		const mockQueryRunner = {
 			connect: jest.fn(),
 			startTransaction: jest.fn(),
 			commitTransaction: jest.fn(),
@@ -89,89 +82,62 @@ describe("PostService", () => {
 				save: jest.fn(),
 				getRepository: jest.fn().mockReturnValue({
 					save: jest.fn(),
-				}),
+				  }),
 			},
 		};
 		mockDataSource.createQueryRunner.mockReturnValue(mockQueryRunner);
 
-		const mockUser = { id: 1, name: "Author" };
-		const mockPostId = 1;
+		const mockCreatePostDto: CreatePostDto = {
+			title: "title",
+			content: "content",
+			doFilter: false,
+			authorId: mockUserId,
+		};
 
+		it("게시물 생성 성공 시 postId를 반환한다.", async () => {
+			mockQueryRunner.manager.save.mockResolvedValue({
+				id: mockPostId,
+			});
 
-    const mockCreatePostDto: CreatePostDto = {
-      title: "title",
-      content: "content",
-      doFilter: false,
-      authorId: mockUser.id,
-    };
-    it("게시물 생성 성공 시 postId를 반환한다.", async () => {
-      jest.spyOn(mockUserRepository, "findOne").mockResolvedValue(
-        mockUser
-      );
-      jest.spyOn(mockQueryRunner.manager, "save").mockResolvedValue({
-        id: mockPostId,
-      });
+			const result = await postService.createPost(mockCreatePostDto);
 
-      const result = await postService.createPost(mockCreatePostDto);
+			expect(mockQueryRunner.connect).toHaveBeenCalled();
+			expect(mockQueryRunner.startTransaction).toHaveBeenCalled();
+			expect(mockQueryRunner.commitTransaction).toHaveBeenCalled();
+			expect(mockQueryRunner.release).toHaveBeenCalled();
 
-      expect(mockQueryRunner.connect).toHaveBeenCalled();
-      expect(mockQueryRunner.startTransaction).toHaveBeenCalled();
-      expect(mockQueryRunner.commitTransaction).toHaveBeenCalled();
-      expect(mockQueryRunner.release).toHaveBeenCalled();
+			expect(result).toBe(mockPostId);
+		});
+		
+		it("게시물 생성 중 오류 발생 시 롤백된다.", async () => {
+			const mockError = new Error("트랜잭션 중 에러 발생");
+			jest.spyOn(mockQueryRunner.manager, "save").mockResolvedValue({
+				id: mockPostId,
+			});
+			mockQueryRunner.manager.getRepository().save.mockRejectedValue(mockError);
 
-      expect(userRepository.findOne).toHaveBeenCalledWith({
-        where: { id: 1 },
-      });
-      expect(result).toBe(mockPostId);
-    });
-    it("존재하지 않는 user일 시 에러를 반환한다.", async () => {
-      jest.spyOn(mockUserRepository, "findOne").mockResolvedValue(
-        null
-      );
+			await expect(
+				postService.createPost(mockCreatePostDto)
+			).rejects.toThrow(mockError);
 
-      await expect(
-        postService.createPost(mockCreatePostDto)
-      ).rejects.toThrow(
-        ServerError.notFound("존재하지 않는 유저입니다")
-      );
-      expect(
-        mockQueryRunner.rollbackTransaction
-      ).not.toHaveBeenCalled();
-      expect(mockQueryRunner.release).toHaveBeenCalled();
-    });
-
-    it("게시물 생성 중 오류 발생 시 롤백된다.", async () => {
-      const mockError = new Error("트랜잭션 중 에러 발생");
-      jest.spyOn(mockUserRepository, "findOne").mockResolvedValue(
-        mockUser
-      );
-      jest.spyOn(mockQueryRunner.manager, "save").mockRejectedValue(
-        mockError
-      );
-
-      await expect(
-        postService.createPost(mockCreatePostDto)
-      ).rejects.toThrow(mockError);
-
-      expect(mockQueryRunner.connect).toHaveBeenCalled();
-      expect(mockQueryRunner.rollbackTransaction).toHaveBeenCalled();
-      expect(mockQueryRunner.release).toHaveBeenCalled();
-    });
+			expect(mockQueryRunner.connect).toHaveBeenCalled();
+			expect(mockQueryRunner.rollbackTransaction).toHaveBeenCalled();
+			expect(mockQueryRunner.release).toHaveBeenCalled();
+		});
 	});
 
 	describe("findPostHeaders", () => {
 		const mockReadPostsQueryDto = {};
 		it("게시물 전체 조회 성공 시 게시물들을 반환한다.", async () => {
-			jest.spyOn(mockPostRepository, "getPostHeaders").mockResolvedValue(
+			mockPostRepository.getPostHeaders.mockResolvedValue(
 				{}
 			);
-
 			const result = await postService.findPostHeaders(
 				mockReadPostsQueryDto,
 				mockUserId
 			);
 
-			expect(postRepository.getPostHeaders).toHaveBeenCalledWith(
+			expect(mockPostRepository.getPostHeaders).toHaveBeenCalledWith(
 				mockReadPostsQueryDto,
 				mockUserId
 			);
@@ -192,26 +158,23 @@ describe("PostService", () => {
 		const mockReadPostsQueryDto = {};
 		it("게시물 개수 조회 성공 시 개수를 반환한다.", async () => {
 			const mockTotal = 10;
-			jest.spyOn(mockPostRepository, "getPostTotal").mockResolvedValue(
-				mockTotal
-			);
+			mockPostRepository.getPostTotal.mockResolvedValue(mockTotal);
 
 			const result = await postService.findPostTotal(
 				mockReadPostsQueryDto,
 				mockUserId
 			);
 
-			expect(postRepository.getPostHeaders).toHaveBeenCalledWith(
+			expect(mockPostRepository.getPostHeaders).toHaveBeenCalledWith(
 				mockReadPostsQueryDto,
 				mockUserId
 			);
 			expect(result).toEqual(mockTotal);
 		});
+
 		it("게시물 개수 조회 도중 에러 발생 시 에러를 반환한다.", async () => {
 			const mockError = new Error("게시물 개수 조회 중 오류 발생");
-			jest.spyOn(mockPostRepository, "getPostTotal").mockRejectedValue(
-				mockError
-			);
+			mockPostRepository.getPostTotal.mockRejectedValue(mockError);
 
 			await expect(
 				postService.findPostTotal(mockReadPostsQueryDto, mockUserId)
@@ -222,10 +185,10 @@ describe("PostService", () => {
 	describe("findPost", () => {
 		it("게시물 상세 조회 성공 시 게시물을 반환한다.", async () => {
 			const mockPost = {id:1, title: "title", content: "content"}
-			jest.spyOn(mockPostRepository, "getPostHeader").mockResolvedValue(mockPost);
+			mockPostRepository.getPostHeader.mockResolvedValue(mockPost);
 			const result = await postService.findPost(mockPostId, mockUserId);
 
-			expect(postRepository.getPostHeader).toHaveBeenCalledWith(
+			expect(mockPostRepository.getPostHeader).toHaveBeenCalledWith(
 				mockPostId,
 				mockUserId
 			);
@@ -233,7 +196,7 @@ describe("PostService", () => {
 		});
 		it("게시물 상세 조회 중 에러 발생 시 에러를 반환한다", async () => {
 			const mockError = new Error("게시물 상세 조회 중 오류 발생");
-			jest.spyOn(mockPostRepository, "getPostHeader").mockRejectedValue(
+			mockPostRepository.getPostHeader.mockRejectedValue(
 				mockError
 			);
 
@@ -250,18 +213,14 @@ describe("PostService", () => {
 			mockUpdateDto = {
 				title: "New Title",
 				content: "New Content",
-				authorId: mockUserId,
+				authorId: mockUser.id,
 				doFilter: true,
 			};
 		});
 		it("게시물 업데이트 성공", async () => {
-			jest.spyOn(mockUserRepository, "findOne").mockResolvedValue(
-				mockUserId
-			);
-			jest.spyOn(mockPostRepository, "findOne").mockResolvedValue(
-				mockPostId
-			);
-			jest.spyOn(mockPostRepository, "update").mockResolvedValue({
+
+			mockPostRepository.findOne.mockResolvedValue(mockPostId);
+			mockPostRepository.update.mockResolvedValue({
 				affected: 1,
 			});
 
@@ -269,33 +228,22 @@ describe("PostService", () => {
 				mockPostId,
 				mockUpdateDto
 			);
-			expect(userRepository.findOne).toHaveBeenCalledWith({
-				where: { id: mockUserId },
-			});
+
 			expect(postRepository.findOne).toHaveBeenCalledWith({
 				where: { id: mockPostId },
 			});
 			expect(postRepository.update).toHaveBeenCalledWith(
-				{ id: mockPostId },
+				{ id: mockPostId, author: {id: mockUserId} },
 				{ title: "New Title", content: "New Content" }
 			);
 			expect(result).toEqual(true);
 		});
 
-		it("존재하지 않는 user 시 에러를 반환한다", async () => {
-			jest.spyOn(mockUserRepository, "findOne").mockResolvedValue(null);
-
-			await expect(
-				postService.updatePost(mockPostId, mockUpdateDto)
-			).rejects.toThrow(ServerError.notFound("없는 유저이거나 존재하지 않는 게시물입니다."));
-
-			expect(postRepository.update).not.toHaveBeenCalled();
-		});
 		it("존재하지 않는 post 시 에러를 반환한다", async () => {
-			jest.spyOn(mockPostRepository, "findOne").mockResolvedValue(null);
+			mockPostRepository.findOne.mockResolvedValue(null);
 			await expect(
 				postService.updatePost(mockPostId, mockUpdateDto)
-			).rejects.toThrow(ServerError.notFound("없는 유저이거나 존재하지 않는 게시물입니다."));
+			).rejects.toThrow(ServerError.notFound("게시물 수정 실패: 존재하지 않는 게시물입니다."));
 
 			expect(postRepository.update).not.toHaveBeenCalled();
 		});
@@ -305,20 +253,20 @@ describe("PostService", () => {
 			jest.clearAllMocks();
 		});
 		it("게시물 삭제 성공", async () => {
-			jest.spyOn(mockPostRepository, "findOne").mockResolvedValue(
+			mockPostRepository.findOne.mockResolvedValue(
 				mockPostId
 			);
-			jest.spyOn(mockPostRepository, "delete").mockResolvedValue({
+			mockPostRepository.delete.mockResolvedValue({
 				affected: 1,
 			});
 
-			const result = await postService.deletePost(mockUserId, mockPostId);
+			const result = await postService.deletePost({authorId: mockUserId, postId: mockPostId});
 
 			expect(postRepository.findOne).toHaveBeenCalledWith({
 				where: { id: mockPostId },
 			});
 			expect(postRepository.update).toHaveBeenCalledWith(
-				{ id: mockPostId, isDelete: 0, author: mockUserId },
+				{ id: mockPostId, isDelete: 0, author: {id : mockUserId} },
 				{ isDelete: 1 }
 			);
 			expect(result).toEqual(true);
@@ -335,8 +283,8 @@ describe("PostService", () => {
 			);
 
 			await expect(
-				postService.deletePost(mockPostId, mockUserId)
-			).rejects.toThrow(ServerError.reference("게시글 삭제 실패"));
+				postService.deletePost({authorId: mockUserId, postId: mockPostId})
+			).rejects.toThrow(ServerError.notFound("게시글 삭제 실패: 존재하지 않는 게시물입니다."));
 
 			expect(postRepository.update).not.toHaveBeenCalled();
 		});
